requirements.txt
# SR Impact Navigator+ 의존성 패키지

# 웹 프레임워크
streamlit>=1.28.0

# Azure 서비스
azure-search-documents>=11.4.0
azure-ai-textanalytics>=5.3.0
azure-identity>=1.15.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# AI/ML
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0

# 시각화
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# 개발 도구
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
